Migrate fjs/effects/mock to .f.mjs, split type-level API into types.ts
Leaf module with no proof.f.ts of its own. Updates 4 dependents
(effects/node/virtual, effects/memory, protocol/mcp, emergent_testing)
to import runtime values from .f.mjs and types from types.ts
separately.
