+++ fjs/effects/memory/proof.f.ts
@@ -1,5 +1,6 @@
 import { assert, assertEq } from '../../asserts/module.f.mjs'
-import { run, type MemOperationMap } from '../mock/module.f.ts'
+import { run } from '../mock/module.f.mjs'
+import type { MemOperationMap } from '../mock/types.ts'
 import { pure, step } from '../module.f.mjs'
 import {
     asBase, asNominal,
