+++ fjs/effects/mock/types.ts
@@ -1,9 +1,9 @@
 /**
- * Mock effect runtimes for testing effectful programs.
+ * Types for mock effect runtimes.
  *
  * @module
  */
-import { match } from "../module.f.mjs"
+
 import type { Effect, Operation, Pr } from "../types.ts"
 
 /**
@@ -19,22 +19,3 @@ export type RunInstance<O extends Operation, S> =
     (state: S) =>
     <O1 extends O, T>(effect: Effect<O1, T>) =>
     readonly[S, T]
-
-export const run =
-    <O extends Operation, S>(o: MemOperationMap<O, S>): RunInstance<O, S> =>
-    state =>
-    effect =>
-{
-    const next = match(o)
-    let s = state
-    let e = effect
-    while (true) {
-        const r = next(e)
-        if (r[0] === 'done') {
-            return [s, r[1]]
-        }
-        const [ns, m] = r[1](s)
-        s = ns
-        e = r[2](m)
-    }
-}
