+++ fjs/effects/node/virtual/module.f.ts
@@ -9,7 +9,8 @@ import { utf8ToString } from '../../../text/module.f.mjs'
 import type { Vec } from '../../../types/bit_vec/types.ts'
 import { empty, length, maxLengthBytes, msb, vec } from '../../../types/bit_vec/module.f.mjs'
 import { error, ok } from '../../../types/result/module.f.mjs'
-import { run, type MemOperationMap, type RunInstance } from '../../mock/module.f.ts'
+import { run } from '../../mock/module.f.mjs'
+import type { MemOperationMap, RunInstance } from '../../mock/types.ts'
 import { asBase, asNominal } from '../../memory/module.f.mjs'
 import type { Key } from '../../memory/types.ts'
 import type { Dirent, FileStat, IoResult, Module, NodeOp, NodeProgramOptions, SandboxResult } from '../module.f.ts'
