+++ fjs/protocol/mcp/proof.f.ts
@@ -2,7 +2,8 @@ import { assert, assertEq } from '../../asserts/module.f.mjs'
 import { pure, step } from '../../effects/module.f.mjs'
 import { eff } from '../../effects/eff/module.f.ts'
 import type { Effect, Operation } from '../../effects/types.ts'
-import { run, type MemOperationMap } from '../../effects/mock/module.f.ts'
+import { run } from '../../effects/mock/module.f.mjs'
+import type { MemOperationMap } from '../../effects/mock/types.ts'
 import { asBase, asNominal, create, read } from '../../effects/memory/module.f.mjs'
 import type { Key, MemOp } from '../../effects/memory/types.ts'
 import type { Unknown } from '../../media/json/module.f.ts'
